+++ changelog.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [0.1.2] - 2024.4.13
+
+### Added
+* get cell value as datetime and time
+
+### Changed
+* output error message in English
+
+
 ## [0.1.1] - 2024.4.13
 
 ### Added